1)deleted and moved files 2)implemented models.py 3)implemented views.py 4)created a superuser     admin     admin@uwm.edu     admin123
