--- CapstoneProject595/user.py
@@ -1,9 +1,0 @@
-class User:
-
-    def __init__(self, name=None, height=0, weight=0, age=0, sex=0):
-        self.name = name
-        self.height = height
-        self.weight = weight
-        self.age = age
-        self.sex = sex
-
