+++ templates/base.html
@@ -8,9 +8,9 @@
     <!-- Bootstrap CSS -->
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css" integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
 {#    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">#}
-    <title>TA Scheduler</title>
+    <title>Exercise website</title>
       <nav class="navbar navbar-expand-sm navbar-dark bg-info flex-nowrap" > <!--style="background-color: #e3f2fd"; -->
-            <span class="navbar-brand w-100">TA scheduler</span>
+            <span class="navbar-brand w-100">Exercise website</span>
             <!-- Navbar content -->
             <div class="navbar-collapse  justify-content-end">
                 <ul class="navbar-nav mx-auto text-white">
